Move callback to callback folder
