+++ smac/callback/__init__.py
@@ -1,7 +1,9 @@
 from smac.callback.callback import Callback
 from smac.callback.metadata_callback import MetadataCallback
+from smac.callback.wandb_logging import WandBCallback
 
 __all__ = [
     "Callback",
     "MetadataCallback",
+    "WandBCallback",
 ]
